Detection: surface file creates so LaunchDaemon/sudoers drops fire (#299)
Subscribe to ES_EVENT_TYPE_NOTIFY_CREATE and re-emit it as an `open` event with synthetic write-mode flags. Creating a NEW file fires NOTIFY_CREATE, not NOTIFY_OPEN, so write-to-sensitive-path rules (privilege_launchd_plist_write, sudoers_tamper) previously missed the canonical new-file-drop attack end-to-end -- the rule + unit tests were correct, but the event never reached the server. Reusing the `open` type means no server/wire change and future write-rules need no extension re-release.

Split handleOpen + handleCreate into ESFSubscriber+FileEvents.swift to keep ESFSubscriber.swift under SwiftLint's file_length/type_body_length caps (same pattern as CDHashHex.swift); serializer is now internal.

Touches ESF subscriptions: must be VM-validated (L5 attack-runbook rule 6) before RC.
